requirements.txt
flask
cs50
requests
pandas
pymongo
python-dotenv
bson.objectid
werkseug.utils